requirements.txt
Flask
flask-swagger-ui
requests
PyAudio
flask_cors
pyyaml
fuzzywuzzy[speedup]
webrtcvad
